Build Transfer File (#7)
This script generates the CSV transfer file that will be used to send 0.1 xDAI to all the gchain allocation recipients.

Unfortunately, at this time the safe-airdrop is not in a state to facilitate back-end scripting. We are working on making this a UI feature, but it may take longer to prepare. In the meantime, I have put together a quick and simple script to facilitate our needs that relies on existing, known and tested transaction building code